requirements.txt
alembic==1.11.1
altair==5.0.1
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.6.2
cachetools==5.3.1
catboost==1.2
category-encoders==2.6.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmaes==0.9.1
colorama==0.4.6
colorlog==6.7.0
colour==0.1.5
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
dtreeviz==2.2.1
exceptiongroup==1.1.1
executing==1.2.0
fonttools==4.40.0
gitdb==4.0.10
GitPython==3.1.31
graphviz==0.20.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipython==8.12.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
lightgbm==3.3.5
llvmlite==0.40.1
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mljar-supervised==1.0.0
numba==0.57.1
numpy==1.23.0
optuna==3.2.0
packaging==23.1
pandas==2.0.2
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
plotly==5.15.0
pluggy==1.2.0
prompt-toolkit==3.0.38
protobuf==4.23.3
pure-eval==0.2.2
pyarrow==12.0.1
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.1.0
pyrsistent==0.19.3
pytest==7.4.0
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.31.0
rich==13.4.2
scikit-learn==1.2.2
scikit-plot==0.3.7
scipy==1.10.1
seaborn==0.12.2
shap==0.41.0
shapely==2.0.1
six==1.16.0
slicer==0.0.7
smmap==5.0.0
SQLAlchemy==2.0.17
stack-data==0.6.2
statsmodels==0.14.0
streamlit==1.24.0
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
wordcloud==1.9.2
xgboost==1.7.6
zipp==3.15.0
